+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `Felt::InputGroup::CheckboxField` component.
 - `Felt::InputGroup::PasswordField` component.
 - `Felt::InputGroup::EmailField` component.
 - `Felt::InputGroup::TextField` component.